move 'cwd' into its own file
Trying to preemptively stem the creeping expansion of dependencies of the
`disppath` “unit” test.
