+++ clink/src/path.c
@@ -1,26 +1,11 @@
 #include "path.h"
 #include "option.h"
 #include <ctype.h>
-#include <errno.h>
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include <sys/stat.h>
-#include <unistd.h>
-
-int cwd(char **wd) {
-
-  if (wd == NULL)
-    return EINVAL;
-
-  char *w = getcwd(NULL, 0);
-  if (w == NULL)
-    return errno;
-
-  *wd = w;
-  return 0;
-}
 
 /// does this path have the given file extension?
 static bool has_ext(const char *path, const char *ext) {
