+++ test/CMakeLists.txt
@@ -22,9 +22,8 @@ add_executable(unit-tests
   dirname.c
   ../../clink/src/dirname.c
   disppath.c
+  ../../clink/src/cwd.c
   ../../clink/src/disppath.c
-  ../../clink/src/option.c
-  ../../clink/src/path.c
   is_root.c
   ../../clink/src/is_root.c
   join.c
